+++ tests/js-custom-authorization/custom_authorization.py
@@ -770,10 +770,6 @@ def test_accept_header(network, args):
         assert r.status_code == http.HTTPStatus.OK.value
         assert r.headers["content-type"] == "application/json"
 
-        r = c.get("/node/commit", headers={"accept": "application/msgpack"})
-        assert r.status_code == http.HTTPStatus.OK.value
-        assert r.headers["content-type"] == "application/msgpack"
-
         r = c.get(
             "/node/commit",
             headers={"accept": "text/html;q=0.9,image/jpeg;video/mpeg;q=0.8,*/*;q=0.1"},
@@ -788,15 +784,6 @@ def test_accept_header(network, args):
         )
         assert r.status_code == http.HTTPStatus.OK.value
         assert r.headers["content-type"] == "application/json"
-
-        r = c.get(
-            "/node/commit",
-            headers={
-                "accept": "text/html;q=0.9,image/jpeg;video/mpeg;q=0.8,application/msgpack;q=0.1"
-            },
-        )
-        assert r.status_code == http.HTTPStatus.OK.value
-        assert r.headers["content-type"] == "application/msgpack"
 
     return network
 
@@ -1310,28 +1297,28 @@ def run_interpreter_reuse(args):
 if __name__ == "__main__":
     cr = ConcurrentRunner()
 
-    cr.add(
-        "authz",
-        run,
-        nodes=infra.e2e_args.nodes(cr.args, 1),
-        js_app_bundle=os.path.join(cr.args.js_app_bundle, "js-custom-authorization"),
-    )
+    # cr.add(
+    #     "authz",
+    #     run,
+    #     nodes=infra.e2e_args.nodes(cr.args, 1),
+    #     js_app_bundle=os.path.join(cr.args.js_app_bundle, "js-custom-authorization"),
+    # )
 
-    cr.add(
-        "limits",
-        run_limits,
-        nodes=infra.e2e_args.nodes(cr.args, 1),
-        js_app_bundle=os.path.join(cr.args.js_app_bundle, "js-limits"),
-    )
+    # cr.add(
+    #     "limits",
+    #     run_limits,
+    #     nodes=infra.e2e_args.nodes(cr.args, 1),
+    #     js_app_bundle=os.path.join(cr.args.js_app_bundle, "js-limits"),
+    # )
 
-    cr.add(
-        "authn",
-        run_authn,
-        nodes=infra.e2e_args.nodes(cr.args, 1),
-        js_app_bundle=os.path.join(cr.args.js_app_bundle, "js-authentication"),
-        initial_user_count=4,
-        initial_member_count=2,
-    )
+    # cr.add(
+    #     "authn",
+    #     run_authn,
+    #     nodes=infra.e2e_args.nodes(cr.args, 1),
+    #     js_app_bundle=os.path.join(cr.args.js_app_bundle, "js-authentication"),
+    #     initial_user_count=4,
+    #     initial_member_count=2,
+    # )
 
     cr.add(
         "content_types",
@@ -1340,18 +1327,18 @@ if __name__ == "__main__":
         js_app_bundle=os.path.join(cr.args.js_app_bundle, "js-content-types"),
     )
 
-    cr.add(
-        "api",
-        run_api,
-        nodes=infra.e2e_args.nodes(cr.args, 1),
-        js_app_bundle=os.path.join(cr.args.js_app_bundle, "js-api"),
-    )
+    # cr.add(
+    #     "api",
+    #     run_api,
+    #     nodes=infra.e2e_args.nodes(cr.args, 1),
+    #     js_app_bundle=os.path.join(cr.args.js_app_bundle, "js-api"),
+    # )
 
-    cr.add(
-        "interpreter_reuse",
-        run_interpreter_reuse,
-        nodes=infra.e2e_args.min_nodes(cr.args, f=1),
-        js_app_bundle=os.path.join(cr.args.js_app_bundle, "js-interpreter-reuse"),
-    )
+    # cr.add(
+    #     "interpreter_reuse",
+    #     run_interpreter_reuse,
+    #     nodes=infra.e2e_args.min_nodes(cr.args, f=1),
+    #     js_app_bundle=os.path.join(cr.args.js_app_bundle, "js-interpreter-reuse"),
+    # )
 
     cr.run()